% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cwa.R
\name{convert_cwa}
\alias{convert_cwa}
\alias{read_cwa_csv}
\title{Convert a CWA activity file to a CSV}
\usage{
convert_cwa(file, outfile = tempfile(fileext = ".csv"), verbose = TRUE)

read_cwa_csv(file)
}
\arguments{
\item{file}{input CWA file}

\item{outfile}{output CSV file}

\item{verbose}{print diagnostic messages}
}
\value{
Name of output CSV file
}
\description{
Convert a CWA activity file to a CSV
}
